Custom URL base and modal for existing images
Added a modal to select and insert existing images. To use it you need to add `"choose-existing"` toolbar option in the field settings.
This will create a custom button in the toolbar that opens the modal. I also changed the image icon to better distinguish between the upload and select image button.

You can now also set a custom url base for the images in the field settings. If set this url will be used instead of the default S3 or whatever url. Helpful for using IMGIX and the like.