Add cache for GraphQL content
Directus only cache content that coming from a GET request, this behavior
works well for REST but GraphQL put the payload of a query in the body
encoded as JSON.

This is a temporary work around, in fact, the cache doesn't consider ACL
and mutation but this really make usable GraphQL integration, because
without this is really slow (above 1.5 second) for simple query.

Consider this a MVP through full GraphQL implementation in Directus.