Fix #1926 (Adds UTC version of created_on field to the payload) (#1987)
* Adds UTC version of created_on field to the payload

This code change fixing suppose to fix the issued described in the following tickets:
https://github.com/directus/api/issues/1926
https://github.com/directus/app/issues/2834
The created_on field is part of the update request for translations data, but its not converted to UTC when update is processing. It's causing TableGateway to fail when updateWith method is called

* Update CoreServicesProvider.php

* Update CoreServicesProvider.php

Added additional validation for existence of "created_on" field in payload and only then make an update.